Include message definitions for communications
This implements a series of message classes that know how to encode
themselves to an endian neutral series of bytes and reconstruct
themselves on the other end.

The network protocol we're going to be using for our TCP communications
requires that every message be preceded by a two byte value that
indicates the length of the actual data, so our encode routines do that
automatically.

For expediency, all classes derive from a base class that gives them
their structure and that base class can defer the decoding of a byte
message to the appropriate class, making the entry point for getting
at a message a single line at the call point.