Include our Connection class
This class abstracts away the core network logic needed to read and send
messages across a TCP connection and raise events for things like
connecting, disconnecting, receiving messages, and so on.

There is an external class, the manager, which is responsible for
actually creating and managing instances of these classes.