Make our panel logging multi-window
For maximum "knowing what the hell is going on" powers, we now log our
panel messages into every window that currently exists, so that no
matter where you are you can open the panel and see the logs.

Utilizing the new API events in builds >= 4050, we detect when a new
window is being created and immediately create a panel in it, copying
the text from one of the other window panels into this one.

This ensures that regardless of the situation at play, every window
always has the full log.