Implement a debounced panel close
Whenever a log goes to the output panel, we now trigger a delayed and
debounced call to close the panel in that window, should it be the
active panel at the end of the delay.

This ensures that we can see events as they happen, with the panel
silently going away as desired.