Don't log to the console and panel simultaneously
This small tweak ensures that if our log is going to the panel, we do
not also log it to the console, since that's a bit redundant.

We will still log to the console for regular logs as well as for error
and message dialogs, since those are more ephemeral.