Include configuration settings for network code
To make our network solution more configurable, this includes a new
settings file and code to obtain the settings using our standard
pattern for this. Also included are command palette and menu entries
for opening the settings.

In order to stop any circular dependencies from occurring, the new
settings function is stored in the utils module.