Allow hostnames in connections for better logs
Previously the Connection class had fields for host and ip to specify
who we are connected to remotely. This was a side effect of the code
being based on the RemoteBuild project from Devember 2018.

In fact all connections we make are to specific IP addresses that are
recovered from Introduction messages, to allow for two hosts that can
talk to each other to always connect even if host names might not
resolve.

For that reason, all instances of host in the Connection class (and all
references to code that used it) now use ip instead, to better signify
what it is that it represents.

In addition a new hostname property was added, which duplicates what the
ip address was initially set to.

Code that makes a connection out immediately sets the hostname to the
value from the Introduction message, and code that accepts an incoming
connection waits for a broadcast Introduction message (always provided
after a successful connect) to set it as well.

This allows logs other than the connect and disconnect to use actual
hostnames, which makes life somewhat easier.