Include new setting for output panel display
This introduces a new setting, `auto_show_panel`, which sets whether the
output panel should automatically be displayed or not when output is
sent to it.

This can be set to `true` to always open the panel, `false` to only open
the panel on errors, and a number to always open the panel, but auto
close it after a set period of time.

When the panel is opened and the value is a boolean, it remains open.