Provide clipboard history syncing
This enabled some clipboard history syncing facilities. Once a new
connection is established, we transmit away our current clipboard
history to the newly connected peer.

This is done via a newly added message that is similar to the clipboard
history message, but which includes information about which entry in
this history this information represents locally, along with the total.

A new setting, `auto_show_panel`, controls this behaviour. When enabled,
we will transmit away and also accept clipboard history. When disabled,
we will not transmit away our history and will actively reject the
history transmitted by the remote end.